renamed to userController.js
